+++ CHANGELOG.md
@@ -5,6 +5,10 @@
   signature of `typing.NewType.__call__`. Patch by Alex Waygood.
 - Fix bug with using `@deprecated` on a mixin class. Inheriting from a
   deprecated class now raises a `DeprecationWarning`. Patch by Jelle Zijlstra.
+- `@deprecated` now gives a better error message if you pass a non-`str`
+  argument to the `msg` parameter. Patch by Alex Waygood.
+- Exclude `__match_args__` from `Protocol` members,
+  this is a backport of https://github.com/python/cpython/pull/110683
 
 # Release 4.8.0 (September 17, 2023)